requirements.txt
pandas==1.5.3
numpy==1.25.2
fuzzywuzzy==0.18.0
scikit-learn==1.2.2
torch==2.1.0
transformers==4.38.2
sqlalchemy==1.4.52
python-dotenv==1.0.1